TalonViewImport updates

+++ v3/TalonViewImport.py
@@ -8,6 +8,9 @@ import shutil
 import glob
 import zipfile
 import csv
+import re
+import pandas
+import datetime
 import onevizion
 
 # Handle command arguments
@@ -87,7 +90,10 @@ Params = ParameterData[ConfigToken]
 onevizion.Config["SMTPToken"] = Params["SMTPToken"]
 
 
-Message = onevizion.TraceMessage
+def Message(Msg,Level=0):
+	"""Prints a message depending on the verbosity level set on the command line"""
+	TraceTag="{TimeStamp}:".format(TimeStamp=datetime.datetime.utcnow().strftime('%Y-%m-%dT%H:%M:%S.%f'))
+	onevizion.TraceMessage(Msg,Level,TraceTag)
 
 # Build up Log for future use if needed
 from collections import OrderedDict
@@ -134,13 +140,58 @@ DocFileFields = Params["COPLineFields"]
 
 # 1) Find files in the Inbound folder with given filemask
 for filename in glob.glob(Params["InboundFolder"]+"/"+Params["FileMask"]):
+	Message("Processing File {File}".format(File=filename),1)
+	#PreProcess CSV file
+	trim = lambda x: x.strip() if type(x) is str else x
+	df = pandas.read_csv(filename).applymap(trim)
+	df.rename(columns=trim, inplace=True)
+	df.to_csv(filename+"2",index=False)
+
+	# find which COP this is
+	CSVData = csv.DictReader(open(filename+"2"))
+
+	COPLIID = ""
+	for row in CSVData:
+		COPLIID = row[Params["COPID-CSV-Col"]]
+		break
+
+	#Get Close Out Package name
+	COP = "-".join(COPLIID.split("-"))[:-1][:-1]
+	Message("This COP is {COP}".format(COP=COP),1)
+
+	# b) unzip document archive
+	COPFolder = Params["InboundFolder"]+"/"+COP
+	if not os.path.exists(COPFolder):
+		os.makedirs(COPFolder)
+
+	# Try to Find ZipFile
+	FileIdent = filename.split("/")[-1].split("_")[0]
+	ZipMask = "{InboundFolder}/{Mask}*.zip".format(InboundFolder=Params["InboundFolder"],Mask=FileIdent)
+	for zfn in glob.glob(ZipMask):
+		ZipFileName = zfn
+		break
+	Message("Unzipping {File}".format(File=ZipFileName),1)
+	#Skip Out if zipfile not found
+	if not os.path.exists(ZipFileName):
+		Message("{FileName} doesn't exist.".format(FileName=ZipFileName))
+		Err=True
+		continue
+	try:
+		#Unzip file
+		with zipfile.ZipFile(ZipFileName,"r") as zip_ref:
+			zip_ref.extractall(COPFolder)
+	except Exception as e:
+		Message("{FileName} can't unzip. Error:\n{Error}".format(FileName=ZipFileName,Error=str(e)))
+		Err=True
+		continue
+
 	# a) Send CSV file to import
 	Message("Sending Import to Trackor for {File}".format(File=filename))
 	Imp = onevizion.Import(
 		paramToken = ConfigToken,
 		impSpecId = Params["ImportID"],
 		action='UPDATE',
-		file = filename
+		file = filename+"2"
 		)
 	if len(Imp.errors) > 0 or Imp.processId is None:
 		Message(filename+" failed to Import from TalonView.")
@@ -149,26 +200,6 @@ for filename in glob.glob(Params["InboundFolder"]+"/"+Params["FileMask"]):
 	else:
 		ProcID = Imp.processId
 		Message("{File} import successfully started. ProcessID={ProcID}".format(File=filename,ProcID=ProcID),1)
-
-	# find which COP this is
-	CSVData = csv.DictReader(open(filename))
-
-	COPLIID = ""
-	for row in CSVData:
-		COPLIID = row[Params["COPID-CSV-Col"]]
-		break
-
-	COP = "-".join(COPLIID.split("-"))[:-1][:-1]
-	Message("This COP is {COP}".format(COP=COP),1)
-
-	# b) unzip document archive
-	COPFolder = Params["InboundFolder"]+"/"+COP
-	if not os.path.exists(COPFolder):
-		os.makedirs(COPFolder)
-	ZipFileName = ".".join(filename.split(".")[:-1])+".zip"
-	Message("Unzipping {File}".format(File=ZipFileName),1)
-	with zipfile.ZipFile(ZipFileName,"r") as zip_ref:
-		zip_ref.extractall(COPFolder)
 
 	# c) Get List of TrackorIDs of COP-Lines with this associated COP as lookup
 	COPLineRequest = onevizion.Trackor(
@@ -192,38 +223,51 @@ for filename in glob.glob(Params["InboundFolder"]+"/"+Params["FileMask"]):
 		# Remove extension
 		Message("Processing File {File}".format(File=docname),1)
 
+		# DocData is the FileName with the folder part and extension stripped off
 		DocData = ".".join(docname.split("/")[-1].split(".")[:-1])
-		# Remove last hyphen section , which is the file number and join the rest back up
-		COPLine = "-".join(DocData.split("-")[:-1])
+		# Remove last hyphen section , which is the file number and join the rest back up, remove the extra whitespace off the ends
+		COPLine = "-".join(DocData.split("-")[:-1]).strip()
 		# Get only the file number
 		try:
-			FileNum = int(DocData[-1])
+			DocDataSplit = DocData.split("(")
+			if len(DocDataSplit) > 1:
+				FileNum = int(DocData.split("(")[-1].strip(")"))
+			else:
+				FileNum = 1
 		except Exception as e:
-			Message("Invalid File name: {File}".format(docname))
+			Message("Invalid File name: {File}".format(File=docname))
 			Err = True
 			continue
 		if FileNum > len(DocFileFields):
-			Message("Invalid Field Number in file name: {File}".format(docname))
+			Message("Invalid Field Number in file name: {File}".format(File=docname))
 			Err = True
 			continue
 
 		# 1) Match TrackorID
 		# 2) send file to Trackor Location for COP-Line and Field
-		COPLineRequest.UploadFile(
-			trackorId = COPLines[COPLine],
-			fieldName = DocFileFields[FileNum - 1],
-			fileName = docname
-			)
+		try:
+			COPLineRequest.UploadFile(
+				trackorId = COPLines[COPLine],
+				fieldName = DocFileFields[FileNum - 1],
+				fileName = docname
+				)
+		except Exception as e:
+			Message("File Upload Failed: {File}\n{Error}".format(File=docname,Error=str(e)))
+			Err = True
+			continue
 
 	# Archive Files so they won't get picked up next time. And Clean up unzipped files
 	ArchiveFolder = Params["InboundFolder"]+"/Archive"
 	if not os.path.exists(ArchiveFolder):
 		os.makedirs(ArchiveFolder)
 	shutil.move(filename,ArchiveFolder)
-	shutil.move(ZipFileName,ArchiveFolder)
+	if os.path.exists(ZipFileName):
+		shutil.move(ZipFileName,ArchiveFolder)
 	shutil.rmtree(Params["InboundFolder"]+"/"+COP)
+	os.remove(filename+"2")
 
 if Err:
 	ErrorNotif("TalonView Import had errors.")
+
